revisi level

+++ app/Http/Controllers/AktaNotarisController.php
@@ -44,10 +44,12 @@ class AktaNotarisController extends Controller
             ->addColumn('action', function ($data) {
                
                 $action = '';
+                $id_berkas = $data->berkas->id_user;
+                if ($id_berkas ==  Auth::user()->id_user ||  Auth::user()->level_user == 'Admin' || Auth::user()->level_user == 'Superadmin') {
                 $action .= "<a href='" . route('akta-notaris.detail', $data->id_aktanotaris) ."' class='btn btn-icon btn-success'><i class='fa fa-eye'></i></a>&nbsp;";
                 $action .= "<a href='javascript:void(0)' class='btn btn-icon btn-primary' data-id='{$data->id_aktanotaris}' onclick='showAktaNotaris(this);'><i class='fa fa-edit'></i></a>&nbsp;";
                 $action .= "<a href='javascript:void(0)' class='btn btn-icon btn-danger'  data-id='{$data->id_aktanotaris}' onclick='deleteAktaNotaris(this);'><i class='fa fa-trash'></i></a>&nbsp;";
-
+                }
                 return $action;
             })
             ->addColumn('tanggal', function ($data) {
